code

+++ setup_db.py
@@ -8,7 +8,8 @@ TEMP_DATA_DIR_NAME = 'opentrain_data'
 TAR_FILE = 'data/2013.tar.gz'
 
 def extract_tar(tar_file, path):
-  shutil.rmtree(path)
+  if os.path.isdir(path):
+    shutil.rmtree(path)
   os.mkdir(path)
   print "Extracting data files to {}".format(path)
   tar = tarfile.open(tar_file)
@@ -22,4 +23,4 @@ src.manage.create_db()
 for filename in os.listdir(temp_data_dir):
   full_filename = os.path.join(temp_data_dir, filename)
   src.manage.add_opentrain_data(full_filename)
-
+shutil.rmtree(temp_data_dir)
